fix: process merged list of all Gas and Elec meter reads rather than separate lists
This change is intended to ensure that Gas and Elec reads (for dual-fuel customers)
are submitted to Ensek as close together as possible. This is required because once
the first read is submitted to Ensek, there is a 10 minute window before a bill is
generated for the customer. If the second read is not received in this window, an
estimated read will be used.