Merged in feature/DMRE-1633-fix-pandas (pull request #38)
DMRE-1633: fix issues resulting from Pandas upgrade

* DMRE-1633: Use explicit column ordering when writing CSV files

When writing CSV files, it is necessary to use a fixed column ordering in
order for data to be ingested in to Redshift correctly. After upgrading
pandas from 0.23.4 to 1.1.4 it appears that the default column order has
changed, so we now need to supply a fixed column order when calling to_csv.

* test: fix some unit test breakage

* fix: preserve existing behaviour after upgrading pandas

Pandas 0.25.0 introduced a new parameter 'max_level' to the json_normalize
function. This needs to be set to zero to preserve existing behaviour
when normalizing Weather Data. Without this, the nested 'weather' attribute
is split out in to 3 separate fields which breaks our data ingest.

* fix: add go_cardless column orders

* fix: re-comment previously commented code

* fix: restore previous LIMIT values

Approved-by: Jonathan Steel