Merged in feature/MTRG-452 (pull request #100)
fix: Identify when Elective HH trial accounts have no HH readings

* fix: Identify when Elective HH trial accounts have no HH readings

Update the Elective HH ETL such that it fetches a list containing all
AccountIDs and associated MPxNs that are involved in the trial. This
list is then iterated over when building the D0379. This allows the
ETL code to identify if any AccountID/MPxN combinations are missing
HH data completely for the period in question.

The AccountID / MPxN list is now retrieved from vw_elective_hh_customers.

MTRG-452

Approved-by: Jonathan Steel
Approved-by: Louis Quaintance