Merged in feature/MTRG-583 (pull request #111)
fix: Actual/Estimated indicator in D0379 being set incorrectly

* fix: Actual/Estimated indicator in D0379 being set incorrectly

The Actua/Estimated indicator is currently being set to the value of
measurement class, however it should be set to 'A', 'C' or 'E'
appropriately.

As we only ever include actual reads as received from the meter, we
should always set the A/E indicator to 'A'.

MTRG-583

Approved-by: Jonathan Steel