Merged in feature/MTRG-689 (pull request #129)
feat: update DAG failure slack alert to optionally include some task documentation

* feat: update DAG failure slack alert to optionally include some task documentation

This change updates the Slack alert message sent on DAG failure such that if the
failing task has a .doc_md or .doc string attached then it is included in the
Slack message.

As part of this change, I’ve implemented a test DAG that we can use to experiment
with failure handling / documentation etc.

MTRG-689

Approved-by: Jonathan Steel
Approved-by: Simon Hunt