Merged in feature/MTRG-1371 (pull request #9)
fix: provision read access to prod S3 buckets for UAT / Preprod s3-sync users

* fix: provision read access to prod S3 buckets for UAT / Preprod s3-sync users

Update permissions for the s3-sync user - this is the IAM user that most ETLs
currently run as. It needs read / write permission to the CDW S3 buckets in
the current AWS account, and read-only permissino to the CDW S3 buckets in
the prod account.

MTRG-1371


Approved-by: Rob Gill
Approved-by: Ross Burchnall
Approved-by: Ryan Pearce